+++ include/shark/Algorithms/DirectSearch/Operators/Domination/ParetoDominance.h
@@ -28,8 +28,8 @@
  * along with Shark.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-#ifndef SHARK_ALGORITHMS_DIRECTSEARCH_PARETODOMINANCE_H
-#define SHARK_ALGORITHMS_DIRECTSEARCH_PARETODOMINANCE_H
+#ifndef SHARK_ALGORITHMS_DIRECTSEARCH_OPERATORS_DOMINATION_PARETODOMINANCE_H
+#define SHARK_ALGORITHMS_DIRECTSEARCH_OPERATORS_DOMINATION_PARETODOMINANCE_H
 
 
 #include <shark/LinAlg/Base.h>
