+++ include/shark/Algorithms/DirectSearch/Operators/Hypervolume/HypervolumeApproximator.h
@@ -33,7 +33,7 @@
 #define HYPERVOLUME_APPROXIMATOR_H
 
 #include <shark/Algorithms/DirectSearch/FitnessExtractor.h>
-#include <shark/Algorithms/DirectSearch/ParetoDominance.h>
+#include <shark/Algorithms/DirectSearch/Operators/Domination/ParetoDominance.h>
 #include <shark/Statistics/Distributions/MultiNomialDistribution.h>
 
 #include <shark/LinAlg/Base.h>
